removing compilation error
Former-commit-id: dbb2fc24ded84837b535926e403ee86751a03c43

+++ java/src/xmipp/viewer/particlepicker/training/gui/SupervisedPickerJFrame.java
@@ -185,7 +185,7 @@ public class SupervisedPickerJFrame extends ParticlePickerJFrame {
         if (new File(dir).isDirectory()) {
             //System.err.println("JM_DEBUG: ============= import from Folder ============");
             result = ppicker.importParticlesFromFolder(dir, format, preffix, suffix, scale, invertx, inverty);
-            boolean resize = ((int)sizetf.getValue()) != ppicker.getSize();
+            boolean resize = ((Integer)sizetf.getValue()).intValue() != ppicker.getSize();
             sizetf.setValue(ppicker.getSize());
             getCanvas().repaint();
             updateMicrographsModel(true);
